ManagerFragment: Try a queuing method for uninstalling on Samsung

+++ app/src/main/java/projekt/substratum/fragments/ManagerFragment.java
@@ -25,6 +25,7 @@ import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Handler;
@@ -93,6 +94,7 @@ import static projekt.substratum.util.files.MapUtils.sortMapByValues;
 public class ManagerFragment extends Fragment implements SearchView.OnQueryTextListener {
 
     private static final int MANAGER_FRAGMENT_INITIAL_DELAY = 500;
+    private static final int UNINSTALL_REQUEST_CODE = 24763851;
     private ArrayList<String> activated_overlays;
     private ManagerAdapter mAdapter;
     private MaterialSheetFab materialSheetFab;
@@ -115,6 +117,7 @@ public class ManagerFragment extends Fragment implements SearchView.OnQueryTextL
     private SearchView searchView;
     private String userInput = "";
     private Boolean first_boot = true;
+    private ArrayList<String> queuedUninstall;
 
     private void resetRecyclerView() {
         // Initialize the recycler view with an empty adapter first
@@ -401,6 +404,26 @@ public class ManagerFragment extends Fragment implements SearchView.OnQueryTextL
             new LayoutReloader(ManagerFragment.this, userInput).execute();
         }
         return true;
+    }
+
+    private void uninstallMultipleAPKS() {
+        if (queuedUninstall.size() > 0) {
+            Uri packageURI = Uri.parse("package:" + queuedUninstall.get(0));
+            Intent uninstallIntent = new Intent(Intent.ACTION_DELETE, packageURI);
+            getActivity().startActivityForResult(uninstallIntent, UNINSTALL_REQUEST_CODE);
+        }
+
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        switch (requestCode) {
+            case UNINSTALL_REQUEST_CODE:
+                if (queuedUninstall != null && queuedUninstall.size() > 0) {
+                    queuedUninstall.remove(0);
+                    uninstallMultipleAPKS();
+                }
+        }
     }
 
     private static class LayoutReloader extends AsyncTask<Void, Void, Void> {
@@ -994,6 +1017,7 @@ public class ManagerFragment extends Fragment implements SearchView.OnQueryTextL
         protected void onPreExecute() {
             ManagerFragment fragment = ref.get();
             if (fragment != null) {
+                fragment.queuedUninstall = new ArrayList<>();
                 fragment.materialSheetFab.hideSheet();
                 fragment.loadingBar.setVisibility(View.VISIBLE);
             }
@@ -1011,15 +1035,18 @@ public class ManagerFragment extends Fragment implements SearchView.OnQueryTextL
                 for (int i = 0; i < len; i++) {
                     ManagerItem overlay1 = fragment.overlayList.get(i);
                     if (overlay1.isSelected()) data.add(overlay1.getName());
+                    if (Systems.isSamsungDevice(context))
+                        fragment.queuedUninstall.add(overlay1.getName());
                 }
 
                 // The magic goes here
                 if (!data.isEmpty()) {
-                    ThemeManager.uninstallOverlay(context, data);
-
+                    if (!Systems.isSamsungDevice(context)) {
+                        ThemeManager.uninstallOverlay(context, data);
+                    }
                     if (!Systems.checkThemeInterfacer(context) &&
                             Packages.needsRecreate(context, data) &&
-                            !Systems.isSamsung(context)) {
+                            !Systems.isSamsungDevice(context)) {
                         Handler handler = new Handler(Looper.getMainLooper());
                         handler.postDelayed(() -> {
                             // OMS may not have written all the changes so quickly just yet
@@ -1056,7 +1083,11 @@ public class ManagerFragment extends Fragment implements SearchView.OnQueryTextL
         protected void onPostExecute(Void result) {
             ManagerFragment fragment = ref.get();
             if (fragment != null) {
-                new LayoutReloader(fragment, fragment.userInput).execute();
+                if (!Systems.isSamsungDevice(fragment.context)) {
+                    new LayoutReloader(fragment, fragment.userInput).execute();
+                } else {
+                    fragment.uninstallMultipleAPKS();
+                }
             }
         }
     }
